requirements.txt
# PostgreSQL MCP Server Dependencies

# Core MCP and database dependencies
mcp>=1.0.0
asyncpg>=0.30.0
python-dotenv>=1.1.1

# Development and testing dependencies
pytest>=8.4.1
pytest-asyncio>=1.0.0

# Optional: For enhanced logging and monitoring
structlog>=24.0.0
